a little działa

+++ core/src/pl/space_marine/game/impediments/Enemy.java
@@ -11,7 +11,7 @@ public abstract class Enemy extends Impediment {
     protected List<Bullet> bullets;
     protected float bulletDamage;
 
-    public Enemy(Image image, Listener listener, float collisionDamage, float direction, int speed, int x, int y) {
+    public Enemy(Image image, Listener listener, float collisionDamage, int direction, int speed, int x, int y) {
         super(image, listener, collisionDamage, direction, speed, x, y);
     }
